[Fix] - 스키마와 엔티티 사이의 컬럼 타입 불일치로 인한 버그 해결 (#549)

+++ backend/src/main/java/kr/touroot/travelogue/domain/TravelogueCountry.java
@@ -37,9 +37,9 @@ public class TravelogueCountry {
     private CountryCode countryCode;
 
     @Column(nullable = false)
-    private Long count;
+    private Integer count;
 
-    public TravelogueCountry(Travelogue travelogue, CountryCode countryCode, Long count) {
+    public TravelogueCountry(Travelogue travelogue, CountryCode countryCode, Integer count) {
         this.travelogue = travelogue;
         this.countryCode = countryCode;
         this.count = count;
